requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
aiohttp==3.10.10
python-dotenv==1.0.1
yookassa==3.10.1
